add modules to files
