+++ index.html
@@ -11,10 +11,10 @@
   <div id="main-title"><h1>Adopt Me</h1></div>
 
   <div id="pet-btns" class="pet-box">
-    <button type="button" class="btn btn-primary">Cats</button>
-    <button type="button" class="btn btn-success">Dogs</button>
-    <button type="button" class="btn btn-danger">Dinos</button>
-    <button type="button" class="btn btn-warning">All Pets</button>
+    <button type="button" class="btn btn-primary" id="cats">Cats</button>
+    <button type="button" class="btn btn-success" id="dogs">Dogs</button>
+    <button type="button" class="btn btn-danger" id="dinos">Dinos</button>
+    <button type="button" class="btn btn-warning" id="all-pets">All Pets</button>
  </div> 
 
  <div id="pet-cards" class="pet-info">
@@ -27,10 +27,13 @@
 
 
 
-  <script src="main.js"></script> 
+
   <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-
+<script src="./javascripts/components/events.js" type="module"></script>
+<script src="./javascripts/components/pets.js" type="module"></script>
+<script src="./javascripts/helpers/util.js" type="module"></script>
+<script src="./javascripts/main.js" type="module"></script>
 </body>
 </html>
