+++ javascripts/helpers/util.js
@@ -1,0 +1,9 @@
+
+// Print to Dom Functions
+const printToDom = (stringToPrint,divId) => {
+  const selectedDiv = document.getElementById(divId);
+  selectedDiv.innerHTML += stringToPrint;
+  };
+
+
+  export{printToDom};
